+++ books-service/src/main/java/com/dayannn/RSOI2/booksservice/BooksServiceApplication.java
@@ -1,0 +1,12 @@
+package com.dayannn.RSOI2.booksservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BooksServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BooksServiceApplication.class, args);
+	}
+}
