+++ gateway-service/src/main/java/com/dayannn/RSOI2/gatewayservice/GatewayServiceApplication.java
@@ -1,0 +1,12 @@
+package com.dayannn.RSOI2.gatewayservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GatewayServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GatewayServiceApplication.class, args);
+	}
+}
