+++ reviews-service/src/main/java/com/dayannn/RSOI2/reviewsservice/ReviewsServiceApplication.java
@@ -1,0 +1,12 @@
+package com.dayannn.RSOI2.reviewsservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReviewsServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReviewsServiceApplication.class, args);
+	}
+}
