--- src/main/java/com/dayannn/RSOI2/Rsoi2Application.java
@@ -1,12 +1,0 @@
-package com.dayannn.RSOI2;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Rsoi2Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Rsoi2Application.class, args);
-	}
-}
