--- src/test/java/com/dayannn/RSOI2/Rsoi2ApplicationTests.java
@@ -1,16 +1,0 @@
-package com.dayannn.RSOI2;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class Rsoi2ApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
