+++ users-service/src/main/java/com/dayannn/RSOI2/usersservice/UsersServiceApplication.java
@@ -1,0 +1,12 @@
+package com.dayannn.RSOI2.usersservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UsersServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UsersServiceApplication.class, args);
+	}
+}
