+++ users-service/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+server.port=15152
+spring.datasource.url=jdbc:postgresql://localhost:5432/server
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.username=program
+spring.datasource.password=test
+
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults = false
+spring.jpa.hibernate.ddl-auto = update
