+++ users-service/src/test/java/com/dayannn/RSOI2/usersservice/UsersServiceApplicationTests.java
@@ -1,0 +1,16 @@
+package com.dayannn.RSOI2.usersservice;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class UsersServiceApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
